feat: MVC패턴으로 리팩터링
