+++ src/main/java/domain/CarRace.java
@@ -1,9 +1,7 @@
-package CarRace;
+package domain;
 
-import java.util.Scanner;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-
 public class CarRace {
 
     public int generateRandom0to9() {
@@ -31,13 +29,6 @@ public class CarRace {
                 .collect(Collectors.joining(", "));
     }
 
-    public void printStatus(int[] carscores, String[] carnames) {
-        for (int i = 0; i < carscores.length; i++) {
-            System.out.println(carnames[i] + " : " + getTrace(carscores[i]));
-        }
-        System.out.println();
-    }
-
     public String getTrace(int carscore) {
         StringBuilder sb = new StringBuilder();
         for (int i = 0; i < carscore; i++) {
@@ -54,26 +45,7 @@ public class CarRace {
         }
     }
 
-    public static void main(String[] args) {
-        CarRace game = new CarRace();
-        game.run();
-    }
 
-    public void run() {
-        Scanner input = new Scanner(System.in);
-        System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
-        String[] carnames = input.nextLine().split(",");
-        validateCarNames(carnames);
 
-        System.out.println("시도할 회수는 몇회인가요?");
-        int n = input.nextInt();
-        int[] carscores = new int[carnames.length];
 
-        System.out.println("\n실행 결과");
-        for (int i = 0; i < n; i++) {
-            playRound(carscores);
-            printStatus(carscores, carnames);
-        }
-        System.out.println(getWinner(carscores, carnames) + "가 최종 우승했습니다.");
-    }
 }
